refactor: order domain - rename and move files
