--- src/main/java/kr/hhplus/be/server/order/interfaces/gateway/SpringDataOrderRepo.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.order.interfaces.gateway;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import kr.hhplus.be.server.order.OrderEntity;
-
-public interface SpringDataOrderRepo extends JpaRepository<OrderEntity, Long> {
-}
